refactor install commands
